Backend/requirements.txt
albumentations
alibi
captum
flask-cors
flask-pymongo
gunicorn
pip-chill
rq
seaborn
spacy-lookups-data
torchvision
